:sparkles: debug: Debug overlay adjustments
